+++ resources/js/components/widgets/SeoProWidget.vue
@@ -10,9 +10,9 @@ defineProps({
 </script>
 
 <template>
-	<Widget :title="__('SEO Pro')" :icon>
+	<Widget :title="__('SEO Pro')" :href="reportsUrl" :icon>
 		<template #actions>
-			<Button :href="reportsUrl" size="sm" :text="__('seo-pro::messages.reports')" />
+			<Button :href="reportsUrl" size="sm" :text="__('seo-pro::messages.view_reports')" />
 		</template>
 
 		<div v-if="report" class="flex flex-col items-center justify-center gap-4" style="min-height: 159px">