+++ src/Fieldtypes/SeoProPreviews.php
@@ -5,13 +5,20 @@ namespace Statamic\SeoPro\Fieldtypes;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Str;
+use Statamic\Contracts\Entries\Collection;
+use Statamic\Contracts\Entries\Entry;
+use Statamic\Contracts\Taxonomies\Taxonomy;
+use Statamic\Contracts\Taxonomies\Term;
 use Statamic\Facades\Antlers;
 use Statamic\Facades\AssetContainer;
 use Statamic\Facades\Site;
 use Statamic\Fields\Fieldtype;
+use Statamic\SeoPro\Fieldtypes\Concerns\ResolvesPlaceholders;
 
 class SeoProPreviews extends Fieldtype
 {
+    use ResolvesPlaceholders;
+
     public $selectable = false;
 
     public function preload()
@@ -24,6 +31,7 @@ class SeoProPreviews extends Fieldtype
                 ? AssetContainer::find(config('statamic.seo-pro.assets.container'))->url()
                 : null,
             'faviconUrl' => $this->faviconUrl(),
+            'placeholders' => $this->getPlaceholders(),
         ];
     }
 
@@ -32,9 +40,9 @@ class SeoProPreviews extends Fieldtype
         $item = $this->field->parent();
 
         $route = match (true) {
-            $item instanceof \Statamic\Contracts\Entries\Entry => $item->route(),
-            $item instanceof \Statamic\Contracts\Entries\Collection => $item->route(Site::selected()->handle()),
-            $item instanceof \Statamic\Contracts\Taxonomies\Taxonomy, $item instanceof \Statamic\Contracts\Taxonomies\Term => '{{ slug }}',
+            $item instanceof Entry => $item->route(),
+            $item instanceof Collection => $item->route(Site::selected()->handle()),
+            $item instanceof Taxonomy, $item instanceof Term => '{{ slug }}',
             default => '',
         };