+++ src/Http/Controllers/CP/PreviewController.php
@@ -3,9 +3,13 @@
 namespace Statamic\SeoPro\Http\Controllers\CP;
 
 use Illuminate\Http\Request;
+use Statamic\Facades\Asset;
 use Statamic\Facades\Blink;
 use Statamic\Facades\Data;
+use Statamic\Facades\Image;
 use Statamic\Http\Controllers\CP\CpController;
+use Statamic\Statamic;
+use Statamic\Support\Arr;
 
 class PreviewController extends CpController
 {
@@ -27,6 +31,25 @@ class PreviewController extends CpController
 
         return [
             'url' => $item->absoluteUrl(),
+            'og_image_url' => $request->image ? $this->transformImage('seo_pro_og', $request->image) : null,
+            'twitter_image_url' => $request->image ? $this->transformImage('seo_pro_twitter', $request->image) : null,
         ];
+    }
+
+    private function transformImage(string $preset, string $assetId): ?string
+    {
+        $asset = Asset::find($assetId);
+
+        if (array_key_exists($preset, Image::customManipulationPresets())) {
+            $glide = Statamic::tag('glide:generate')
+                ->src($asset)
+                ->preset($preset)
+                ->absolute(true)
+                ->fetch();
+
+            return Arr::get($glide, '0.url');
+        }
+
+        return $asset->absoluteUrl();
     }
 }