+++ src/ServiceProvider.php
@@ -14,6 +14,10 @@ use Statamic\Facades\Permission;
 use Statamic\Facades\Site;
 use Statamic\Facades\User;
 use Statamic\Providers\AddonServiceProvider;
+use Statamic\SeoPro\GraphQL\AlternateLocaleType;
+use Statamic\SeoPro\GraphQL\SeoProType;
+use Statamic\SeoPro\Reporting\Page;
+use Statamic\SeoPro\Reporting\Report;
 use Statamic\SeoPro\SiteDefaults\SiteDefaults;
 
 class ServiceProvider extends AddonServiceProvider
@@ -30,6 +34,14 @@ class ServiceProvider extends AddonServiceProvider
     ];
 
     protected $config = false;
+
+    public function register()
+    {
+        $this->registerSerializableClasses([
+            Page::class,
+            Report::class,
+        ]);
+    }
 
     public function bootAddon()
     {
@@ -148,8 +160,8 @@ class ServiceProvider extends AddonServiceProvider
 
     protected function bootAddonGraphQL()
     {
-        GraphQL::addType(\Statamic\SeoPro\GraphQL\SeoProType::class);
-        GraphQL::addType(\Statamic\SeoPro\GraphQL\AlternateLocaleType::class);
+        GraphQL::addType(SeoProType::class);
+        GraphQL::addType(AlternateLocaleType::class);
 
         $seoField = function () {
             return [